User permissions 1

+++ plugins/users-manager/controllers/delete-controller.php
@@ -7,10 +7,16 @@
 		$csrf = csrf_verify($postdata);
 		if($csrf){
 			
-			$user->delete($row->id);
-			
-			message_success("Registro eliminado exitosamente!");
-			redirect($admin_route . '/' . $plugin_route);
+			if(user_can('delete_user')){
+				
+				$user->delete($row->id);
+				
+				if(file_exists($row->image))
+					unlink($row->image);
+				
+				message_success("Registro eliminado exitosamente!");
+				redirect($admin_route . '/' . $plugin_route);
+			}
 		}
 		
 		$user->errors['email'] = "El formulario expiró!";
